rename: User Notice Retrieve merge into Notice Retreive
