+++ src/main/java/com/liberty52/auth/service/applicationservice/NoticeRetrieveService.java
@@ -6,6 +6,10 @@ import org.springframework.data.domain.Pageable;
 
 public interface NoticeRetrieveService {
 
+    NoticeRetrieveResponse retrieveUserNotice(Pageable pageable);
+
+    NoticeDetailResponse retrieveUserNoticeDetail(String noticeId);
+
     NoticeRetrieveResponse retrieveNoticesByAdmin(String role, Pageable pageable);
 
     NoticeDetailResponse retrieveNoticeDetailByAdmin(String role, String noticeId);
