+++ src/main/java/com/liberty52/auth/service/applicationservice/impl/NoticeRetrieveServiceImpl.java
@@ -10,15 +10,53 @@ import com.liberty52.auth.service.entity.Notice;
 import com.liberty52.auth.service.repository.NoticeRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
+import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Pageable;
+import org.springframework.data.domain.Sort;
 import org.springframework.stereotype.Service;
 import org.springframework.util.CollectionUtils;
+
+import static com.liberty52.auth.global.utils.PagingUtils.*;
 
 @Service
 @RequiredArgsConstructor
 public class NoticeRetrieveServiceImpl implements NoticeRetrieveService {
 
     private final NoticeRepository noticeRepository;
+
+    @Override
+    public NoticeRetrieveResponse retrieveUserNotice(Pageable pageable) {
+
+        int pageNumber = pageable.getPageNumber();
+        pageable = PageRequest.of(pageNumber, pageable.getPageSize(),
+                Sort.by(CREATED_AT).descending());
+
+        Page<Notice> notices = noticeRepository.findAll(pageable);
+        int totalPages = notices.getTotalPages();
+        validatePageNumber(pageNumber, totalPages);
+        return createResponse(pageable, pageNumber, notices, totalPages);
+
+    }
+
+    @Override
+    public NoticeDetailResponse retrieveUserNoticeDetail(String noticeId) {
+        return new NoticeDetailResponse(noticeRepository.findById(noticeId)
+                .orElseThrow(() ->
+                        new NoticeNotFoundById(noticeId)));
+    }
+
+    private NoticeRetrieveResponse createResponse(Pageable pageable, int pageNumber,
+                                                  Page<Notice> notices, int totalPages) {
+        if (isEmptyPageNum(pageable, totalPages))
+            return NoticeRetrieveResponse.EMPTY_RESOURCE_RESPONSE;
+
+        return new NoticeRetrieveResponse(notices.getContent(),
+                getPageInfo(totalPages, pageNumber));
+    }
+
+    private boolean isEmptyPageNum(Pageable pageable, int totalPages) {
+        return totalPages == 0 && pageable.getPageNumber() == 0;
+    }
 
     @Override
     public NoticeRetrieveResponse retrieveNoticesByAdmin(String role, Pageable pageable) {
@@ -42,4 +80,6 @@ public class NoticeRetrieveServiceImpl implements NoticeRetrieveService {
 
         return new NoticeDetailResponse(notice);
     }
+
+
 }
